pyproject.toml
[project]
name = "cflow"
version = "0.1.0"
description = "Machine First Functional Graph Workflow Builder"
authors = [
    { name = "James Ott", email = "jlgott@users.noreply.github.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.95.1",
    "openai-agents>=0.1.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
